CM-1894 Refactor mci/alerts stuff into their own files.
Trying to make things a bit more manageable/readable/testable.
